sdk/js: move fs dependent methods
